Bug 1912695: Clean up uses of offset() and offset64(). r=yury,rhunt
This patch attempts to clean up and clarify code surrounding offsets in
baseline and ion. No behavior was changed, except in the case of
removing some unnecessary conversions.

Differential Revision: https://phabricator.services.mozilla.com/D219931

UltraBlame original commit: 4fa0b44a5701d38c6f6dc6d1af1a34f26bb7a194

+++ js/src/wasm/WasmMemory.h
@@ -160,7 +160,7 @@ extern Pages ClampedMaxPages(IndexType t, Pages initialPages,
 
 extern size_t ComputeMappedSize(Pages clampedMaxPages);
 
-extern size_t GetMaxOffsetGuardLimit(bool hugeMemory);
+extern uint64_t GetMaxOffsetGuardLimit(bool hugeMemory);
 
 
 extern bool IsValidBoundsCheckImmediate(uint32_t i);
